Main

Documentation

main :: IO ()

The core IO part of the code

contains :: [String] -> [Bool] -> [String]

Takes a list of words, and a list of Booleans and outputs the words corresponding to the the True values in the list of Booleans

To be used with startsWith, endsWith, and containsWildcard functions

containsElem :: [String] -> [Char] -> [String]

Takes in a list of words and a list of characters, and outputs a subset of the list of words which contain all of the input characters

lessThan :: [String] -> Int -> [String]

Outputs a subset of strings which have lengths less than the input Int, from the input [String]

greaterThan :: [String] -> Int -> [String]

Outputs a subset of strings which have lengths greater than the input Int, from the input [String]

equalTo :: [String] -> Int -> [String]

Outputs a subset of strings which have lengths equal to the input Int, from the input [String]

checkNumber :: [Char] -> Bool

Checks if input String is a Digit

Used for user input selection

startsWith :: [Char] -> [Char] -> Bool

Checks if a string starts with another string

Also utilizes wildcards denoted by "$" in the second argument

containsWildcard :: [Char] -> [Char] -> Bool

Checks if a string contains a list of characters

Also utilizes wildcards denoted by "$" in the second argument

endsWith :: [Char] -> [Char] -> Bool

Checks if a string ends with another string

Also utilizes wildcards denoted by "$" in the second argument